+++ dqops/src/main/java/com/dqops/checks/column/monitoring/anomaly/ColumnAnomalyDailyMonitoringChecksSpec.java
@@ -45,6 +45,8 @@ public class ColumnAnomalyDailyMonitoringChecksSpec extends AbstractCheckCategor
             put("daily_sum_anomaly", o -> o.dailySumAnomaly);
             put("daily_mean_anomaly", o -> o.dailyMeanAnomaly);
             put("daily_median_anomaly", o -> o.dailyMedianAnomaly);
+            put("daily_min_anomaly", o -> o.dailyMinAnomaly);
+            put("daily_max_anomaly", o -> o.dailyMaxAnomaly);
 
             put("daily_mean_change", o -> o.dailyMeanChange);
             put("daily_mean_change_1_day", o -> o.dailyMeanChange1Day);
@@ -71,6 +73,14 @@ public class ColumnAnomalyDailyMonitoringChecksSpec extends AbstractCheckCategor
 
     @JsonPropertyDescription("Verifies that the median in a column changes in a rate within a percentile boundary during the last 90 days.")
     private ColumnMedianAnomalyStationaryCheckSpec dailyMedianAnomaly;
+
+    @JsonPropertyDescription("Detects new outliers, which are new minimum values, much below the last known minimum value. " +
+            "If the minimum value is constantly changing, detects outliers as the biggest change of the minimum value during the last 90 days.")
+    private ColumnMinAnomalyDifferencingCheckSpec dailyMinAnomaly;
+
+    @JsonPropertyDescription("Detects new outliers, which are new maximum values, much above the last known maximum value. " +
+            "If the maximum value is constantly changing, detects outliers as the biggest change of the maximum value during the last 90 days.")
+    private ColumnMaxAnomalyDifferencingCheckSpec dailyMaxAnomaly;
 
     @JsonPropertyDescription("Verifies that the mean value in a column changed in a fixed rate since the last readout.")
     private ColumnMeanChangeCheckSpec dailyMeanChange;
@@ -169,6 +179,42 @@ public class ColumnAnomalyDailyMonitoringChecksSpec extends AbstractCheckCategor
         this.setDirtyIf(!Objects.equals(this.dailyMedianAnomaly, dailyMedianAnomaly));
         this.dailyMedianAnomaly = dailyMedianAnomaly;
         propagateHierarchyIdToField(dailyMedianAnomaly, "daily_median_anomaly");
+    }
+
+    /**
+     * Returns a min anomaly for the last 90 days.
+     * @return Min anomaly for the last 90 days.
+     */
+    public ColumnMinAnomalyDifferencingCheckSpec getDailyMinAnomaly() {
+        return dailyMinAnomaly;
+    }
+
+    /**
+     * Sets a new minimum anomaly for the last 90 days.
+     * @param dailyMinAnomaly New minimum anomaly.
+     */
+    public void setDailyMinAnomaly(ColumnMinAnomalyDifferencingCheckSpec dailyMinAnomaly) {
+        this.setDirtyIf(!Objects.equals(this.dailyMinAnomaly, dailyMinAnomaly));
+        this.dailyMinAnomaly = dailyMinAnomaly;
+        propagateHierarchyIdToField(dailyMinAnomaly, "daily_min_anomaly");
+    }
+
+    /**
+     * Returns a max anomaly for the last 90 days.
+     * @return Max anomaly for the last 90 days.
+     */
+    public ColumnMaxAnomalyDifferencingCheckSpec getDailyMaxAnomaly() {
+        return dailyMaxAnomaly;
+    }
+
+    /**
+     * Sets a new maximum anomaly for the last 90 days.
+     * @param dailyMaxAnomaly New maximum anomaly.
+     */
+    public void setDailyMaxAnomaly(ColumnMaxAnomalyDifferencingCheckSpec dailyMaxAnomaly) {
+        this.setDirtyIf(!Objects.equals(this.dailyMaxAnomaly, dailyMaxAnomaly));
+        this.dailyMaxAnomaly = dailyMaxAnomaly;
+        propagateHierarchyIdToField(dailyMaxAnomaly, "daily_max_anomaly");
     }
 
     /**